requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
discord.py==2.3.1
distlib==0.3.6
filelock==3.12.2
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
platformdirs==3.8.1
python-dotenv==0.15.0
requests==2.31.0
urllib3==2.0.3
virtualenv==20.23.1
yarl==1.9.2
